refactor workbench performance details route
